+++ src/components/Form/utility/index.tsx
@@ -78,35 +78,3 @@ export const FieldCounter = ({ value = "", maxLength = MAX_SHORT_TEXT_LENGTH, cl
 export const FieldHelperText = ({ message = "" }: { message: React.ReactNode }) => (
   <span className="field-helper-text">{typeof message === "string" ? addLineBreaks(message) : message}</span>
 );
-
-export const FieldReloadMessage = ({
-  message = '',
-  onReload,
-  reloaderStatus
-}: {
-  message?: string
-  onReload?: () => void
-  reloaderStatus?: Enum<typeof RELOADSTATUS>
-}) => {
-  if (!reloaderStatus || reloaderStatus === RELOADSTATUS.SUCCESS) {
-    return null
-  }
-  return (
-    <div className="field-error-reloader">
-      {
-        {
-          LOADING: <Spinner />,
-          FAILED: (
-            <>
-              <i className="notification-icon icon-ab-exclamation-mark-filled" />
-              {` ${message} `}
-              <Button appearance="link" onClick={() => onReload && onReload()}>
-                {translation('common.fieldReloadMessage.reload')}
-              </Button>
-            </>
-          )
-        }[reloaderStatus]
-      }
-    </div>
-  )
-}