+++ src/utils/typography.tsx
@@ -13,5 +13,3 @@ export const addLineBreaks = (text: string, marginTop: string | number = 4) => {
     );
   });
 };
-
-export const trimExcessiveSpace = (text: string) => text.replace(/\s{2,}/g, ' ').trim()