+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.10.4]
+### Added
+- `dataQaProps` for options in `StyledRadioTabVertical`
+
 ## [0.10.3]
 ### Added
 - styling for right icon props in valid field text